test(insights): move persons-modal e2e to its own spec
CI's "verify changed Playwright tests are stable" step reruns every test in a changed file 10x. Adding the persons-modal test to trends.spec.ts pulled a pre-existing flaky test ("Save an insight…", which intermittently times out waiting on the save button) into that gate, failing the check for reasons unrelated to this PR.

Move the persons-modal navigation test into its own spec so flake verification only exercises this new test, and trends.spec.ts returns to its master state. The shared InsightPage locators are unchanged.

Generated-By: PostHog Code
Task-Id: 48f57754-412c-4dd2-b15d-7db1df9c7cee
